+++ scenedetect/detectors/threshold_detector.py
@@ -18,7 +18,7 @@ This detector is available from the command-line as the `detect-threshold` comma
 
 from enum import Enum
 from logging import getLogger
-from typing import List, Optional
+import typing as ty
 
 import numpy
 
@@ -114,10 +114,11 @@ class ThresholdDetector(SceneDetector):
         }
         self._metric_keys = [ThresholdDetector.THRESHOLD_VALUE_KEY]
 
-    def get_metrics(self) -> List[str]:
+    @property
+    def metric_keys(self) -> ty.List[str]:
         return self._metric_keys
 
-    def process_frame(self, frame_num: int, frame_img: numpy.ndarray) -> List[int]:
+    def process_frame(self, frame_num: int, frame_img: numpy.ndarray) -> ty.List[int]:
         """Process the next frame. `frame_num` is assumed to be sequential.
 
         Args: