+++ scenedetect/scene_detector.py
@@ -25,6 +25,7 @@ in order to be compatible with PySceneDetect.
     event (in, out, cut, etc...).
 """
 
+from abc import ABC, abstractmethod
 from enum import Enum
 import typing as ty
 
@@ -33,60 +34,37 @@ import numpy
 from scenedetect.stats_manager import StatsManager
 
 
-# pylint: disable=unused-argument, no-self-use
-class SceneDetector:
-    """ Base class to inherit from when implementing a scene detection algorithm.
+class SceneDetector(ABC):
+    """Base class to inherit from when implementing a scene detection algorithm.
 
-    This API is not yet stable and subject to change.
+    This API is not yet stable and subject to change. Currently has a very simple interface, where
+    on each frame, a detector emits a list of points where scene cuts are detected.
 
-    This represents a "dense" scene detector, which returns a list of frames where
-    the next scene/shot begins in a video.
-
-    Also see the implemented scene detectors in the scenedetect.detectors module
-    to get an idea of how a particular detector can be created.
+    Also see the implemented scene detectors in the scenedetect.detectors module to get an idea of
+    how a particular detector can be created. In the future, this will be changed to support
+    different types of detections (e.g. fades versus cuts) and confidence scores of each event.
     """
-    # TODO(v0.7): Make this a proper abstract base class.
 
-    stats_manager: ty.Optional[StatsManager] = None
-    """Optional :class:`StatsManager <scenedetect.stats_manager.StatsManager>` to
-    use for caching frame metrics to and from."""
+    def __init__(self):
+        self._stats_manager = None
 
-    # TODO(v1.0): Remove - this is a rarely used case for what is now a neglegible performance gain.
-    def is_processing_required(self, frame_num: int) -> bool:
-        """[DEPRECATED] DO NOT USE
+    @property
+    def stats_manager(self) -> ty.Optional[StatsManager]:
+        """Optional :class:`StatsManager <scenedetect.stats_manager.StatsManager>` to
+        use for caching frame metrics to and from."""
+        return self._stats_manager
 
-        Test if all calculations for a given frame are already done.
+    @stats_manager.setter
+    def stats_manager(self, new_manager):
+        self._stats_manager = new_manager
 
-        Returns:
-            False if the SceneDetector has assigned _metric_keys, and the
-            stats_manager property is set to a valid StatsManager object containing
-            the required frame metrics/calculations for the given frame - thus, not
-            needing the frame to perform scene detection.
+    @property
+    @abstractmethod
+    def metric_keys(self) -> ty.List[str]:
+        """List of all metric names/keys used by the detector."""
+        raise NotImplementedError
 
-            True otherwise (i.e. the frame_img passed to process_frame is required
-            to be passed to process_frame for the given frame_num).
-        """
-        metric_keys = self.get_metrics()
-        return not metric_keys or not (self.stats_manager is not None
-                                       and self.stats_manager.metrics_exist(frame_num, metric_keys))
-
-    def stats_manager_required(self) -> bool:
-        """Stats Manager Required: Prototype indicating if detector requires stats.
-
-        Returns:
-            True if a StatsManager is required for the detector, False otherwise.
-        """
-        return False
-
-    def get_metrics(self) -> ty.List[str]:
-        """Get Metrics:  Get a list of all metric names/keys used by the detector.
-
-        Returns:
-            List of strings of frame metric key names that will be used by
-            the detector when a StatsManager is passed to process_frame.
-        """
-        return []
-
+    @abstractmethod
     def process_frame(self, frame_num: int, frame_img: numpy.ndarray) -> ty.List[int]:
         """Process the next frame. `frame_num` is assumed to be sequential.
 
@@ -102,12 +80,12 @@ class SceneDetector:
         Returns:
             List of frame numbers of cuts to be added to the cutting list.
         """
-        return []
+        raise NotImplementedError
 
     def post_process(self, frame_num: int) -> ty.List[int]:
         """Post Process: Performs any processing after the last frame has been read.
 
-        Prototype method, no actual detection.
+        Default implementation is a no-op.
 
         Returns:
             List of frame numbers of cuts to be added to the cutting list.
@@ -121,16 +99,26 @@ class SceneDetector:
         """
         return 0
 
+    # DEPRECATED METHODS TO BE REMOVED IN v1.0
+
+    def is_processing_required(self, frame_num: int) -> bool:
+        """[DEPRECATED] DO NOT USE"""
+        return True
+
+    def stats_manager_required(self) -> bool:
+        """[DEPRECATED] DO NOT USE"""
+        return False
+
+    def get_metrics(self) -> ty.List[str]:
+        """[DEPRECATED] USE `metric_keys` PROPERTY INSTEAD"""
+        return self.metric_keys
+
 
 class SparseSceneDetector(SceneDetector):
-    """Base class to inherit from when implementing a sparse scene detection algorithm.
+    """[DEPRECATED - DO NOT USE]
 
-    This class will be removed in v1.0 and should not be used.
-
-    Unlike dense detectors, sparse detectors detect "events" and return a *pair* of frames,
-    as opposed to just a single cut.
-
-    An example of a SparseSceneDetector is the MotionDetector.
+    This class will be removed in v1.0, with the goal being the SceneDetector interface will emit
+    event types and confidence scores rather than having different interfaces.
     """
 
     def process_frame(self, frame_num: int,
@@ -157,6 +145,7 @@ class SparseSceneDetector(SceneDetector):
         return []
 
 
+# TODO(v0.7): Add documentation.
 class FlashFilter:
 
     class Mode(Enum):